Move etcd function into own file.
